@@ -37,7 +37,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.5
-  qr_flutter: ^4.0.0
+  qr_flutter: ^4.1.0
   qr_code_scanner: ^1.0.1
   # http: ^0.13.3
   # http_parser: ^4.0.0
